chore<BE> authorization 추가 null 수정

+++ BACKEND/src/main/java/com/ssafy/challympic/service/AdminService.java
@@ -73,8 +73,10 @@ public class AdminService {
     @Transactional
     public void deletePostByChallenge(int challenge_no) {
         List<Post> postList = postRepository.findByChallengNo(challenge_no);
-        for (Post post : postList) {
-            postRepository.deleteByPostNo(post.getPost_no());
+        if(!postList.isEmpty()){
+            for (Post post : postList) {
+                postRepository.deleteByPostNo(post.getPost_no());
+            }
         }
     }
